Add autoload magic comments to files.
Hopefully I have covered all of the needed autoloads. Anything is a bit tricky
since a lot of its functionality is provided through variables, rather than
functions. This makes autoloading a bit more difficult.

This also means that "anything-startup.el" is no longer needed, as it is handled
automagically by the autoloads.

Signed-off-by: Dan Hackney <dan@haxney.org>